Move access control provider to app.tsx
