+++ apps/console/src/app.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2023-2024, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -17,26 +17,31 @@
  */
 
 import { BasicUserInfo, DecodedIDTokenPayload, useAuthContext } from "@asgardeo/auth-react";
+import { AccessControlProvider, AllFeatureInterface, FeatureGateInterface } from "@wso2is/access-control";
 import { AppConstants as CommonAppConstants } from "@wso2is/core/constants";
+import { IdentityAppsApiException } from "@wso2is/core/exceptions";
 import { CommonHelpers, isPortalAccessGranted } from "@wso2is/core/helpers";
 import { RouteInterface, StorageIdentityAppsSettingsInterface, emptyIdentityAppsSettings } from "@wso2is/core/models";
 import { setI18nConfigs, setServiceResourceEndpoints } from "@wso2is/core/store";
 import { AuthenticateUtils, LocalStorageUtils } from "@wso2is/core/utils";
+import useAuthorization from "@wso2is/features/admin.authorization.v1/hooks/use-authorization";
 import { EventPublisher, PreLoader } from "@wso2is/features/admin.core.v1";
 import { ProtectedRoute } from "@wso2is/features/admin.core.v1/components";
 import { Config, DocumentationLinks, getBaseRoutes } from "@wso2is/features/admin.core.v1/configs";
 import { AppConstants } from "@wso2is/features/admin.core.v1/constants";
 import { history } from "@wso2is/features/admin.core.v1/helpers";
 import useResourceEndpoints from "@wso2is/features/admin.core.v1/hooks/use-resource-endpoints";
-import useRoutes from "@wso2is/features/admin.core.v1/hooks/use-routes";
 import {
     ConfigReducerStateInterface,
     DocumentationLinksInterface,
     FeatureConfigInterface,
     ServiceResourceEndpointsInterface
 } from "@wso2is/features/admin.core.v1/models";
-import { AppState } from "@wso2is/features/admin.core.v1/store";
+import { AppState, store } from "@wso2is/features/admin.core.v1/store";
 import { commonConfig } from "@wso2is/features/admin.extensions.v1";
+import { useGetAllFeatures } from "@wso2is/features/admin.extensions.v1/components/feature-gate/api/feature-gate";
+import { featureGateConfig } from "@wso2is/features/admin.extensions.v1/configs/feature-gate";
+import { OrganizationUtils } from "@wso2is/features/admin.organizations.v1/utils";
 import { I18nModuleOptionsInterface } from "@wso2is/i18n";
 import {
     ChunkErrorModal,
@@ -49,6 +54,7 @@ import {
 } from "@wso2is/react-components";
 import has from "lodash-es/has";
 import isEmpty from "lodash-es/isEmpty";
+import set from "lodash-es/set";
 import * as moment from "moment";
 import React, { FunctionComponent, ReactElement, Suspense, useEffect, useState } from "react";
 import { Helmet } from "react-helmet";
@@ -66,14 +72,15 @@ import "moment/locale/fr";
  * @returns App Root component.
  */
 export const App: FunctionComponent<Record<string, never>> = (): ReactElement => {
+    const featureGateConfigUpdated : FeatureGateInterface = { ...featureGateConfig };
 
     const dispatch: Dispatch<any> = useDispatch();
-
-    const { filterRoutes } = useRoutes();
 
     const eventPublisher: EventPublisher = EventPublisher.getInstance();
 
     const { trySignInSilently, getDecodedIDToken, signOut, state } = useAuthContext();
+
+    const { legacyAuthzRuntime }  = useAuthorization();
 
     const { setResourceEndpoints } = useResourceEndpoints();
 
@@ -84,14 +91,19 @@ export const App: FunctionComponent<Record<string, never>> = (): ReactElement =>
     const appTitle: string = useSelector((state: AppState) => state?.config?.ui?.appTitle);
     const uuid: string = useSelector((state: AppState) => state.profile.profileInfo.id);
     const theme: string = useSelector((state: AppState) => state?.config?.ui?.theme?.name);
-    const isFirstLevelOrg: boolean = useSelector(
-        (state: AppState) => state.organization.isFirstLevelOrganization
-    );
     const allowedScopes: string = useSelector((state: AppState) => state?.auth?.allowedScopes);
+    const organizationType: string = useSelector((state: AppState) => state?.organization?.organizationType);
 
     const [ baseRoutes, setBaseRoutes ] = useState<RouteInterface[]>(getBaseRoutes());
     const [ sessionTimedOut, setSessionTimedOut ] = useState<boolean>(false);
-    const [ routesFiltered, setRoutesFiltered ] = useState<boolean>(false);
+    const [ orgId, setOrgId ] = useState<string>();
+    const [ featureGateConfigData, setFeatureGateConfigData ] =
+        useState<FeatureGateInterface | null>(featureGateConfigUpdated);
+
+    const {
+        data: allFeatures,
+        error: featureGateAPIException
+    } = useGetAllFeatures(orgId, state.isAuthenticated);
 
     /**
      * Set the deployment configs in redux state.
@@ -209,12 +221,49 @@ export const App: FunctionComponent<Record<string, never>> = (): ReactElement =>
     }, [ uuid ]);
 
     useEffect(() => {
-        if (!state.isAuthenticated) {
+        if(state.isAuthenticated) {
+            if (OrganizationUtils.isSuperOrganization(store.getState().organization.organization)
+            || store.getState().organization.isFirstLevelOrganization) {
+                getDecodedIDToken().then((response: DecodedIDTokenPayload)=>{
+                    const orgName: string = response.org_name;
+                    // Set org_name instead of org_uuid as the API expects org_name
+                    // as it resolves tenant uuid from it.
+
+                    setOrgId(orgName);
+                });
+            } else {
+                // Set the sub org id to the current organization id.
+                setOrgId(store.getState().organization.organization.id);
+            }
+        }
+    }, [ state ]);
+
+    useEffect(() => {
+        if (allFeatures instanceof IdentityAppsApiException || featureGateAPIException) {
             return;
         }
 
-        filterRoutes(() => setRoutesFiltered(true), isFirstLevelOrg);
-    }, [ filterRoutes, state.isAuthenticated, isFirstLevelOrg ]);
+        if (!allFeatures) {
+            return;
+        }
+
+        if (allFeatures?.length > 0) {
+            allFeatures.forEach((feature: AllFeatureInterface )=> {
+                // converting the identifier to path.
+                const path: string = feature.featureIdentifier.replace(/-/g, ".");
+                // Obtain the status and set it to the feature gate config.
+                const featureStatusPath: string = `${ path }.status`;
+
+                set(featureGateConfigUpdated,featureStatusPath, feature.featureStatus);
+
+                const featureTagPath: string = `${ path }.tags`;
+
+                set(featureGateConfigUpdated,featureTagPath, feature.featureTags);
+
+                setFeatureGateConfigData(featureGateConfigUpdated);
+            });
+        }
+    }, [ allFeatures ]);
 
     /**
      * Set the value of Session Timed Out.
@@ -280,7 +329,7 @@ export const App: FunctionComponent<Record<string, never>> = (): ReactElement =>
             });
     };
 
-    if (!routesFiltered || isEmpty(config?.deployment) || isEmpty(config?.endpoints)) {
+    if (isEmpty(config?.deployment) || isEmpty(config?.endpoints)) {
         return <PreLoader/>;
     }
 
@@ -290,192 +339,199 @@ export const App: FunctionComponent<Record<string, never>> = (): ReactElement =>
                 <DocumentationProvider<DocumentationLinksInterface> links={ DocumentationLinks }>
                     <Suspense fallback={ <PreLoader /> }>
                         <MediaContextProvider>
-                            <SessionManagementProvider
-                                onSessionTimeoutAbort={ handleSessionTimeoutAbort }
-                                onSessionLogout={ handleSessionLogout }
-                                onLoginAgain={ handleStayLoggedIn }
-                                setSessionTimedOut={ handleSessionTimeOut }
-                                sessionTimedOut={ sessionTimedOut }
-                                modalOptions={ {
-                                    description: (
-                                        <Trans
-                                            i18nKey={
-                                                "console:common.modals.sessionTimeoutModal." +
-                                                "description"
-                                            }
-                                        >
-                                            When you click on the <Code>Go back</Code> button, we
-                                            will try to recover the session if it exists. If you
-                                            don&apos;t have an active session, you will be
-                                            redirected to the login page
-                                        </Trans>
-                                    ),
-                                    headingI18nKey: "console:common.modals.sessionTimeoutModal" +
-                                        ".heading",
-                                    loginAgainButtonText: (
-                                        <Trans
-                                            i18nKey={
-                                                "console:common.modals" +
-                                                ".sessionTimeoutModal.loginAgainButton"
-                                            }>
-                                            Login again
-                                        </Trans>
-                                    ),
-                                    primaryButtonText: (
-                                        <Trans
-                                            i18nKey={
-                                                "console:common.modals" +
-                                                ".sessionTimeoutModal.primaryButton"
-                                            }>
-                                            Go back
-                                        </Trans>
-                                    ),
-                                    secondaryButtonText: (
-                                        <Trans
-                                            i18nKey={
-                                                "console:common.modals" +
-                                                ".sessionTimeoutModal.secondaryButton"
-                                            }>
-                                            Logout
-                                        </Trans>
-                                    ),
-                                    sessionTimedOutDescription: (
-                                        <Trans
-                                            i18nKey={
-                                                "console:common.modals" +
-                                                ".sessionTimeoutModal.sessionTimedOutDescription"
-                                            }>
-                                            Please log in again to continue from where you left off.
-                                        </Trans>
-                                    ),
-                                    sessionTimedOutHeadingI18nKey: "console:common.modals" +
-                                        ".sessionTimeoutModal.sessionTimedOutHeading"
-                                } }
-                                type={ SessionTimeoutModalTypes.DEFAULT }
+                            <AccessControlProvider
+                                allowedScopes={ allowedScopes }
+                                features={ featureGateConfigData }
+                                isLegacyRuntimeEnabled={ legacyAuthzRuntime }
+                                organizationType={ organizationType }
                             >
-                                <>
-                                    <Helmet>
-                                        <title>{ appTitle }</title>
-                                        {
-                                            (window?.themeHash && window?.publicPath && theme)
-                                                ? (
-                                                    <link
-                                                        href={
-                                                            `${
-                                                                window?.origin
-                                                            }${
-                                                                window?.publicPath
-                                                            }/libs/themes/${
-                                                                theme
-                                                            }/theme.${ window?.themeHash }.min.css`
-                                                        }
-                                                        rel="stylesheet"
-                                                        type="text/css"
-                                                    />
-                                                )
-                                                : null
-                                        }
-                                    </Helmet>
-                                    <NetworkErrorModal
-                                        heading={
-                                            (<Trans
-                                                i18nKey={ "common:networkErrorMessage.heading" }
-                                            >
-                                                Your session has expired
-                                            </Trans>)
-                                        }
-                                        description={
-                                            (<Trans
-                                                i18nKey={ "common:networkErrorMessage.description" }
-                                            >
-                                                Please try signing in again.
-                                            </Trans>)
-                                        }
-                                        primaryActionText={
-                                            (<Trans
+                                <SessionManagementProvider
+                                    onSessionTimeoutAbort={ handleSessionTimeoutAbort }
+                                    onSessionLogout={ handleSessionLogout }
+                                    onLoginAgain={ handleStayLoggedIn }
+                                    setSessionTimedOut={ handleSessionTimeOut }
+                                    sessionTimedOut={ sessionTimedOut }
+                                    modalOptions={ {
+                                        description: (
+                                            <Trans
                                                 i18nKey={
-                                                    "common:networkErrorMessage.primaryActionText"
+                                                    "console:common.modals.sessionTimeoutModal." +
+                                                    "description"
                                                 }
                                             >
-                                                Sign In
-                                            </Trans>)
-                                        }
-                                        primaryAction={
-                                            signOut
-                                        }
-                                    />
-                                    <ChunkErrorModal
-                                        heading={
-                                            (<Trans
+                                                When you click on the <Code>Go back</Code> button, we
+                                                will try to recover the session if it exists. If you
+                                                don&apos;t have an active session, you will be
+                                                redirected to the login page
+                                            </Trans>
+                                        ),
+                                        headingI18nKey: "console:common.modals.sessionTimeoutModal" +
+                                            ".heading",
+                                        loginAgainButtonText: (
+                                            <Trans
                                                 i18nKey={
-                                                    "common:chunkLoadErrorMessage.heading"
-                                                }
-                                            >
-                                                Something went wrong
-                                            </Trans>)
-                                        }
-                                        description={
-                                            (<Trans
+                                                    "console:common.modals" +
+                                                    ".sessionTimeoutModal.loginAgainButton"
+                                                }>
+                                                Login again
+                                            </Trans>
+                                        ),
+                                        primaryButtonText: (
+                                            <Trans
                                                 i18nKey={
-                                                    "common:chunkLoadErrorMessage.description"
-                                                }
-                                            >
-                                                An error occurred when serving the requested
-                                                application. Please try reloading the app.
-                                            </Trans>)
-                                        }
-                                        primaryActionText={
-                                            (<Trans
+                                                    "console:common.modals" +
+                                                    ".sessionTimeoutModal.primaryButton"
+                                                }>
+                                                Go back
+                                            </Trans>
+                                        ),
+                                        secondaryButtonText: (
+                                            <Trans
                                                 i18nKey={
-                                                    "common:chunkLoadErrorMessage.primaryActionText"
-                                                }
-                                            >
-                                                Reload the App
-                                            </Trans>)
-                                        }
-                                    />
-                                    <Switch>
-                                        <Redirect
-                                            exact
-                                            from="/"
-                                            to={ AppConstants.getAppHomePath() }
+                                                    "console:common.modals" +
+                                                    ".sessionTimeoutModal.secondaryButton"
+                                                }>
+                                                Logout
+                                            </Trans>
+                                        ),
+                                        sessionTimedOutDescription: (
+                                            <Trans
+                                                i18nKey={
+                                                    "console:common.modals" +
+                                                    ".sessionTimeoutModal.sessionTimedOutDescription"
+                                                }>
+                                                Please log in again to continue from where you left off.
+                                            </Trans>
+                                        ),
+                                        sessionTimedOutHeadingI18nKey: "console:common.modals" +
+                                            ".sessionTimeoutModal.sessionTimedOutHeading"
+                                    } }
+                                    type={ SessionTimeoutModalTypes.DEFAULT }
+                                >
+                                    <>
+                                        <Helmet>
+                                            <title>{ appTitle }</title>
+                                            {
+                                                (window?.themeHash && window?.publicPath && theme)
+                                                    ? (
+                                                        <link
+                                                            href={
+                                                                `${
+                                                                    window?.origin
+                                                                }${
+                                                                    window?.publicPath
+                                                                }/libs/themes/${
+                                                                    theme
+                                                                }/theme.${ window?.themeHash }.min.css`
+                                                            }
+                                                            rel="stylesheet"
+                                                            type="text/css"
+                                                        />
+                                                    )
+                                                    : null
+                                            }
+                                        </Helmet>
+                                        <NetworkErrorModal
+                                            heading={
+                                                (<Trans
+                                                    i18nKey={ "common:networkErrorMessage.heading" }
+                                                >
+                                                    Your session has expired
+                                                </Trans>)
+                                            }
+                                            description={
+                                                (<Trans
+                                                    i18nKey={ "common:networkErrorMessage.description" }
+                                                >
+                                                    Please try signing in again.
+                                                </Trans>)
+                                            }
+                                            primaryActionText={
+                                                (<Trans
+                                                    i18nKey={
+                                                        "common:networkErrorMessage.primaryActionText"
+                                                    }
+                                                >
+                                                    Sign In
+                                                </Trans>)
+                                            }
+                                            primaryAction={
+                                                signOut
+                                            }
                                         />
-                                        {
-                                            baseRoutes.map((route: RouteInterface, index: number) => {
-                                                return (
-                                                    route.protected ?
-                                                        (
-                                                            <ProtectedRoute
-                                                                component={ route.component }
-                                                                path={ route.path }
-                                                                key={ index }
-                                                                exact={ route.exact }
-                                                            />
-                                                        )
-                                                        :
-                                                        (
-                                                            <Route
-                                                                path={ route.path }
-                                                                render={
-                                                                    (props:  RouteComponentProps<
-                                                                        { [p: string]: string },
-                                                                        StaticContext, unknown
-                                                                    >) => {
-                                                                        return (<route.component
-                                                                            { ...props }
-                                                                        />);
+                                        <ChunkErrorModal
+                                            heading={
+                                                (<Trans
+                                                    i18nKey={
+                                                        "common:chunkLoadErrorMessage.heading"
+                                                    }
+                                                >
+                                                    Something went wrong
+                                                </Trans>)
+                                            }
+                                            description={
+                                                (<Trans
+                                                    i18nKey={
+                                                        "common:chunkLoadErrorMessage.description"
+                                                    }
+                                                >
+                                                    An error occurred when serving the requested
+                                                    application. Please try reloading the app.
+                                                </Trans>)
+                                            }
+                                            primaryActionText={
+                                                (<Trans
+                                                    i18nKey={
+                                                        "common:chunkLoadErrorMessage.primaryActionText"
+                                                    }
+                                                >
+                                                    Reload the App
+                                                </Trans>)
+                                            }
+                                        />
+                                        <Switch>
+                                            <Redirect
+                                                exact
+                                                from="/"
+                                                to={ AppConstants.getAppHomePath() }
+                                            />
+                                            {
+                                                baseRoutes.map((route: RouteInterface, index: number) => {
+                                                    return (
+                                                        route.protected ?
+                                                            (
+                                                                <ProtectedRoute
+                                                                    component={ route.component }
+                                                                    path={ route.path }
+                                                                    key={ index }
+                                                                    exact={ route.exact }
+                                                                />
+                                                            )
+                                                            :
+                                                            (
+                                                                <Route
+                                                                    path={ route.path }
+                                                                    render={
+                                                                        (props:  RouteComponentProps<
+                                                                            { [p: string]: string },
+                                                                            StaticContext, unknown
+                                                                        >) => {
+                                                                            return (<route.component
+                                                                                { ...props }
+                                                                            />);
+                                                                        }
                                                                     }
-                                                                }
-                                                                key={ index }
-                                                                exact={ route.exact }
-                                                            />
-                                                        )
-                                                );
-                                            })
-                                        }
-                                    </Switch>
-                                </>
-                            </SessionManagementProvider>
+                                                                    key={ index }
+                                                                    exact={ route.exact }
+                                                                />
+                                                            )
+                                                    );
+                                                })
+                                            }
+                                        </Switch>
+                                    </>
+                                </SessionManagementProvider>
+                            </AccessControlProvider>
                         </MediaContextProvider>
                     </Suspense>
                 </DocumentationProvider>
